backend/requirements.txt
# Django Core
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
django-simple-history>=3.3
django-guardian>=2.4

# Database
psycopg2-binary>=2.9
dj-database-url>=2.0

# Environment & Configuration
python-dotenv>=1.0

# Production Server
gunicorn>=21.0

# Background Tasks
celery>=5.3.6
redis>=4.5

# AI Integration
openai>=1.0

# Testing (opcional)
pytest>=7.0
pytest-django>=4.5
factory-boy>=3.2

# Development (opcional)
django-debug-toolbar>=4.0
django-extensions>=3.2

# System monitoring
psutil>=5.9